Change the default integrator to sock Langevin
The splitting integrator is anywhere from 50-80% slower than stock Langevin, so this uses the stock one as default. If we can speed up the splitting integrator, we can change this back in the future.

Also fixes a bug where the default was "BAOAB" which fails since it expects R and V