Allow multiple experiments to keep running if one NaN's
Enables multiple and/or combinatorial experiments to run if one of them goes to NaN.
Introduces a special error which can be trapped and sent up the stack from a sampler through the phase and experiment. Should also print out to the end of the logger a warning to the user if any NaN.

This has NOT been tested yet and should NOT be merged, but I want to get the implementation discussion going while I figure out the correct way to test this.

Fixes #328, Fixes #616