Total rework of the Topography selection engine
The new `Topography.select` function is a workhorse function with its own internal extendable classes for handling selection strings while trying to minimize cyclomatic complextiy. The restraint selection logic has been relegated mostly to the Topography selection.

Because this is mostly a rework of the internal, this should be ready for review.