Make restoration and last iteration more robust
This is the first of a couple changes to make checkpoint restoration and access more robust.

Next step will to be make `MultiStateAnalyzer` able to detect corruption and back up X number of checkpoints. I split the PRs to minimize amount which has to be reviewed and to set the underlying structure first